Move violations for pycodestyle rules to rules files
