+++ src/registry.rs
@@ -13,26 +13,26 @@ use crate::{rules, violations};
 
 ruff_macros::define_rule_mapping!(
     // pycodestyle errors
-    E101 => violations::MixedSpacesAndTabs,
+    E101 => rules::pycodestyle::rules::MixedSpacesAndTabs,
     E401 => violations::MultipleImportsOnOneLine,
     E402 => violations::ModuleImportNotAtTopOfFile,
-    E501 => violations::LineTooLong,
-    E711 => violations::NoneComparison,
-    E712 => violations::TrueFalseComparison,
-    E713 => violations::NotInTest,
-    E714 => violations::NotIsTest,
-    E721 => violations::TypeComparison,
-    E722 => violations::DoNotUseBareExcept,
-    E731 => violations::DoNotAssignLambda,
-    E741 => violations::AmbiguousVariableName,
-    E742 => violations::AmbiguousClassName,
-    E743 => violations::AmbiguousFunctionName,
+    E501 => rules::pycodestyle::rules::LineTooLong,
+    E711 => rules::pycodestyle::rules::NoneComparison,
+    E712 => rules::pycodestyle::rules::TrueFalseComparison,
+    E713 => rules::pycodestyle::rules::NotInTest,
+    E714 => rules::pycodestyle::rules::NotIsTest,
+    E721 => rules::pycodestyle::rules::TypeComparison,
+    E722 => rules::pycodestyle::rules::DoNotUseBareExcept,
+    E731 => rules::pycodestyle::rules::DoNotAssignLambda,
+    E741 => rules::pycodestyle::rules::AmbiguousVariableName,
+    E742 => rules::pycodestyle::rules::AmbiguousClassName,
+    E743 => rules::pycodestyle::rules::AmbiguousFunctionName,
     E902 => violations::IOError,
     E999 => violations::SyntaxError,
     // pycodestyle warnings
-    W292 => violations::NoNewLineAtEndOfFile,
-    W505 => violations::DocLineTooLong,
-    W605 => violations::InvalidEscapeSequence,
+    W292 => rules::pycodestyle::rules::NoNewLineAtEndOfFile,
+    W505 => rules::pycodestyle::rules::DocLineTooLong,
+    W605 => rules::pycodestyle::rules::InvalidEscapeSequence,
     // pyflakes
     F401 => violations::UnusedImport,
     F402 => violations::ImportShadowedByLoopVar,
