+++ src/rules/pycodestyle/rules/ambiguous_class_name.rs
@@ -1,7 +1,21 @@
+use ruff_macros::derive_message_formats;
+
 use crate::ast::types::Range;
+use crate::define_violation;
 use crate::registry::Diagnostic;
 use crate::rules::pycodestyle::helpers::is_ambiguous_name;
-use crate::violations;
+use crate::violation::Violation;
+
+define_violation!(
+    pub struct AmbiguousClassName(pub String);
+);
+impl Violation for AmbiguousClassName {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        let AmbiguousClassName(name) = self;
+        format!("Ambiguous class name: `{name}`")
+    }
+}
 
 /// E742
 pub fn ambiguous_class_name<F>(name: &str, locate: F) -> Option<Diagnostic>
@@ -10,7 +24,7 @@ where
 {
     if is_ambiguous_name(name) {
         Some(Diagnostic::new(
-            violations::AmbiguousClassName(name.to_string()),
+            AmbiguousClassName(name.to_string()),
             locate(),
         ))
     } else {
