+++ src/rules/pycodestyle/rules/do_not_use_bare_except.rs
@@ -1,10 +1,22 @@
+use ruff_macros::derive_message_formats;
 use rustpython_ast::{Excepthandler, Stmt, StmtKind};
 use rustpython_parser::ast::Expr;
 
 use crate::ast::helpers::except_range;
+use crate::define_violation;
 use crate::registry::Diagnostic;
 use crate::source_code::Locator;
-use crate::violations;
+use crate::violation::Violation;
+
+define_violation!(
+    pub struct DoNotUseBareExcept;
+);
+impl Violation for DoNotUseBareExcept {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!("Do not use bare `except`")
+    }
+}
 
 /// E722
 pub fn do_not_use_bare_except(
@@ -19,7 +31,7 @@ pub fn do_not_use_bare_except(
             .any(|stmt| matches!(stmt.node, StmtKind::Raise { exc: None, .. }))
     {
         Some(Diagnostic::new(
-            violations::DoNotUseBareExcept,
+            DoNotUseBareExcept,
             except_range(handler, locator),
         ))
     } else {
