+++ src/rules/pycodestyle/rules/mod.rs
@@ -1,16 +1,16 @@
-pub use ambiguous_class_name::ambiguous_class_name;
-pub use ambiguous_function_name::ambiguous_function_name;
-pub use ambiguous_variable_name::ambiguous_variable_name;
-pub use do_not_assign_lambda::do_not_assign_lambda;
-pub use do_not_use_bare_except::do_not_use_bare_except;
-pub use doc_line_too_long::doc_line_too_long;
-pub use invalid_escape_sequence::invalid_escape_sequence;
-pub use line_too_long::line_too_long;
-pub use literal_comparisons::literal_comparisons;
-pub use mixed_spaces_and_tabs::mixed_spaces_and_tabs;
-pub use no_newline_at_end_of_file::no_newline_at_end_of_file;
-pub use not_tests::not_tests;
-pub use type_comparison::type_comparison;
+pub use ambiguous_class_name::{ambiguous_class_name, AmbiguousClassName};
+pub use ambiguous_function_name::{ambiguous_function_name, AmbiguousFunctionName};
+pub use ambiguous_variable_name::{ambiguous_variable_name, AmbiguousVariableName};
+pub use do_not_assign_lambda::{do_not_assign_lambda, DoNotAssignLambda};
+pub use do_not_use_bare_except::{do_not_use_bare_except, DoNotUseBareExcept};
+pub use doc_line_too_long::{doc_line_too_long, DocLineTooLong};
+pub use invalid_escape_sequence::{invalid_escape_sequence, InvalidEscapeSequence};
+pub use line_too_long::{line_too_long, LineTooLong};
+pub use literal_comparisons::{literal_comparisons, NoneComparison, TrueFalseComparison};
+pub use mixed_spaces_and_tabs::{mixed_spaces_and_tabs, MixedSpacesAndTabs};
+pub use no_newline_at_end_of_file::{no_newline_at_end_of_file, NoNewLineAtEndOfFile};
+pub use not_tests::{not_tests, NotInTest, NotIsTest};
+pub use type_comparison::{type_comparison, TypeComparison};
 
 mod ambiguous_class_name;
 mod ambiguous_function_name;
