+++ src/violations.rs
@@ -38,191 +38,6 @@ impl Violation for ModuleImportNotAtTopOfFile {
 }
 
 define_violation!(
-    pub struct LineTooLong(pub usize, pub usize);
-);
-impl Violation for LineTooLong {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let LineTooLong(length, limit) = self;
-        format!("Line too long ({length} > {limit} characters)")
-    }
-}
-
-define_violation!(
-    pub struct MixedSpacesAndTabs;
-);
-impl Violation for MixedSpacesAndTabs {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Indentation contains mixed spaces and tabs")
-    }
-}
-
-#[derive(Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub enum EqCmpop {
-    Eq,
-    NotEq,
-}
-
-impl From<&Cmpop> for EqCmpop {
-    fn from(cmpop: &Cmpop) -> Self {
-        match cmpop {
-            Cmpop::Eq => EqCmpop::Eq,
-            Cmpop::NotEq => EqCmpop::NotEq,
-            _ => unreachable!("Expected Cmpop::Eq | Cmpop::NotEq"),
-        }
-    }
-}
-
-define_violation!(
-    pub struct NoneComparison(pub EqCmpop);
-);
-impl AlwaysAutofixableViolation for NoneComparison {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let NoneComparison(op) = self;
-        match op {
-            EqCmpop::Eq => format!("Comparison to `None` should be `cond is None`"),
-            EqCmpop::NotEq => format!("Comparison to `None` should be `cond is not None`"),
-        }
-    }
-
-    fn autofix_title(&self) -> String {
-        let NoneComparison(op) = self;
-        match op {
-            EqCmpop::Eq => "Replace with `cond is None`".to_string(),
-            EqCmpop::NotEq => "Replace with `cond is not None`".to_string(),
-        }
-    }
-}
-
-define_violation!(
-    pub struct TrueFalseComparison(pub bool, pub EqCmpop);
-);
-impl AlwaysAutofixableViolation for TrueFalseComparison {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let TrueFalseComparison(value, op) = self;
-        match (value, op) {
-            (true, EqCmpop::Eq) => format!("Comparison to `True` should be `cond is True`"),
-            (true, EqCmpop::NotEq) => {
-                format!("Comparison to `True` should be `cond is not True`")
-            }
-            (false, EqCmpop::Eq) => format!("Comparison to `False` should be `cond is False`"),
-            (false, EqCmpop::NotEq) => {
-                format!("Comparison to `False` should be `cond is not False`")
-            }
-        }
-    }
-
-    fn autofix_title(&self) -> String {
-        let TrueFalseComparison(value, op) = self;
-        match (value, op) {
-            (true, EqCmpop::Eq) => "Replace with `cond is True`".to_string(),
-            (true, EqCmpop::NotEq) => "Replace with `cond is not True`".to_string(),
-            (false, EqCmpop::Eq) => "Replace with `cond is False`".to_string(),
-            (false, EqCmpop::NotEq) => "Replace with `cond is not False`".to_string(),
-        }
-    }
-}
-
-define_violation!(
-    pub struct NotInTest;
-);
-impl AlwaysAutofixableViolation for NotInTest {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Test for membership should be `not in`")
-    }
-
-    fn autofix_title(&self) -> String {
-        "Convert to `not in`".to_string()
-    }
-}
-
-define_violation!(
-    pub struct NotIsTest;
-);
-impl AlwaysAutofixableViolation for NotIsTest {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Test for object identity should be `is not`")
-    }
-
-    fn autofix_title(&self) -> String {
-        "Convert to `is not`".to_string()
-    }
-}
-
-define_violation!(
-    pub struct TypeComparison;
-);
-impl Violation for TypeComparison {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Do not compare types, use `isinstance()`")
-    }
-}
-
-define_violation!(
-    pub struct DoNotUseBareExcept;
-);
-impl Violation for DoNotUseBareExcept {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Do not use bare `except`")
-    }
-}
-
-define_violation!(
-    pub struct DoNotAssignLambda(pub String);
-);
-impl AlwaysAutofixableViolation for DoNotAssignLambda {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Do not assign a `lambda` expression, use a `def`")
-    }
-
-    fn autofix_title(&self) -> String {
-        let DoNotAssignLambda(name) = self;
-        format!("Rewrite `{name}` as a `def`")
-    }
-}
-
-define_violation!(
-    pub struct AmbiguousVariableName(pub String);
-);
-impl Violation for AmbiguousVariableName {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let AmbiguousVariableName(name) = self;
-        format!("Ambiguous variable name: `{name}`")
-    }
-}
-
-define_violation!(
-    pub struct AmbiguousClassName(pub String);
-);
-impl Violation for AmbiguousClassName {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let AmbiguousClassName(name) = self;
-        format!("Ambiguous class name: `{name}`")
-    }
-}
-
-define_violation!(
-    pub struct AmbiguousFunctionName(pub String);
-);
-impl Violation for AmbiguousFunctionName {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let AmbiguousFunctionName(name) = self;
-        format!("Ambiguous function name: `{name}`")
-    }
-}
-
-define_violation!(
     pub struct IOError(pub String);
 );
 impl Violation for IOError {
@@ -241,48 +56,6 @@ impl Violation for SyntaxError {
     fn message(&self) -> String {
         let SyntaxError(message) = self;
         format!("SyntaxError: {message}")
-    }
-}
-
-// pycodestyle warnings
-
-define_violation!(
-    pub struct NoNewLineAtEndOfFile;
-);
-impl AlwaysAutofixableViolation for NoNewLineAtEndOfFile {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("No newline at end of file")
-    }
-
-    fn autofix_title(&self) -> String {
-        "Add trailing newline".to_string()
-    }
-}
-
-define_violation!(
-    pub struct InvalidEscapeSequence(pub char);
-);
-impl AlwaysAutofixableViolation for InvalidEscapeSequence {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let InvalidEscapeSequence(char) = self;
-        format!("Invalid escape sequence: '\\{char}'")
-    }
-
-    fn autofix_title(&self) -> String {
-        "Add backslash to escape sequence".to_string()
-    }
-}
-
-define_violation!(
-    pub struct DocLineTooLong(pub usize, pub usize);
-);
-impl Violation for DocLineTooLong {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let DocLineTooLong(length, limit) = self;
-        format!("Doc line too long ({length} > {limit} characters)")
     }
 }
